Class_2/pyproject.toml
[project]
name = "class-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.68.0",
    "python-dotenv>=1.2.2",
]
